requirements.txt
websockets
python-dotenv
google-auth
requests
